refactor: update import paths to use Web namespace for controllers
