+++ routes/actions.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-use App\Api\Groups\Controllers\GroupClearController;
-use App\Api\Groups\Controllers\GroupToggleController;
-use App\Api\Media\Controllers\MediaDownloadController;
-use App\Api\Notifications\Controllers\MarkAllNotificationsReadController;
-use App\Api\Profiles\Controllers\SwitchProfileController;
-use App\Api\Tags\Controllers\TagOrderController;
-use App\Api\Users\Controllers\UserSettingsController;
-use App\Api\Videos\Controllers\VideoImportController;
-use App\Api\Videos\Controllers\VideoLikeController;
-use App\Api\Videos\Controllers\VideoSaveController;
-use App\Api\Videos\Controllers\VideoTranscodeController;
-use App\Api\Videos\Controllers\VideoTranscodedController;
+use App\Web\Groups\Controllers\GroupClearController;
+use App\Web\Groups\Controllers\GroupToggleController;
+use App\Web\Media\Controllers\MediaDownloadController;
+use App\Web\Notifications\Controllers\MarkAllNotificationsReadController;
+use App\Web\Profiles\Controllers\SwitchProfileController;
+use App\Web\Tags\Controllers\TagOrderController;
+use App\Web\Users\Controllers\UserSettingsController;
+use App\Web\Videos\Controllers\VideoDispatchTranscodeController;
+use App\Web\Videos\Controllers\VideoImportController;
+use App\Web\Videos\Controllers\VideoLikeController;
+use App\Web\Videos\Controllers\VideoSaveController;
+use App\Web\Videos\Controllers\VideoTranscodedController;
 use Illuminate\Support\Facades\Route;
 
 // Settings
@@ -52,6 +52,6 @@ Route::prefix('/videos')->name('videos.')->group(function () {
     Route::post('/import', VideoImportController::class)->name('import');
     Route::post('/{video}/like', VideoLikeController::class)->name('like');
     Route::post('/{video}/save', VideoSaveController::class)->name('save');
-    Route::post('/{video}/transcode', VideoTranscodeController::class)->name('transcode');
+    Route::post('/{video}/transcode', VideoDispatchTranscodeController::class)->name('transcode');
     Route::post('/{video}/transcoded', VideoTranscodedController::class)->name('transcoded');
 });
