+++ src/App/Web/Groups/Controllers/GroupClearController.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace App\Api\Notifications\Controllers;
+namespace App\Web\Groups\Controllers;
 
+use Domain\Groups\Models\Group;
 use Foundation\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
-class MarkAllNotificationsReadController extends Controller implements HasMiddleware
+class GroupClearController extends Controller implements HasMiddleware
 {
     public static function middleware(): array
     {
@@ -22,16 +22,16 @@ class MarkAllNotificationsReadController extends Controller implements HasMiddle
         ];
     }
 
-    public function __invoke(Request $request): RedirectResponse
+    public function __invoke(Group $group): RedirectResponse
     {
-        Gate::authorize('update', $request->user());
+        Gate::authorize('update', $group);
 
-        $request->user()->unreadNotifications()->update(['read_at' => now()]);
+        defer(fn () => $group->videos()->detach());
 
         Inertia::flash([
-            'title' => 'All caught up',
-            'description' => 'All notifications have been marked as read.',
-            'type' => 'success',
+            'title' => $group->title,
+            'description' => __('All videos will be detached shortly.'),
+            'type' => 'info',
         ]);
 
         return back();
