+++ src/App/Web/Groups/Controllers/GroupToggleController.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Api\Videos\Controllers;
+namespace App\Web\Groups\Controllers;
 
-use Domain\Groups\Enums\GroupType;
+use Domain\Groups\Models\Group;
 use Domain\Videos\Models\Video;
 use Foundation\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
@@ -14,7 +14,7 @@ use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
-class VideoSaveController extends Controller implements HasMiddleware
+class GroupToggleController extends Controller implements HasMiddleware
 {
     public static function middleware(): array
     {
@@ -24,18 +24,21 @@ class VideoSaveController extends Controller implements HasMiddleware
         ];
     }
 
-    public function __invoke(Video $video, Request $request): RedirectResponse
+    public function __invoke(Group $group, Video $video, Request $request): RedirectResponse
     {
         Gate::authorize('view', $video);
+        Gate::authorize('update', $group);
 
-        // Toggle the video in the user's saved group.
-        $group = $request->user()->toggleInGroup($video, GroupType::Saved);
+        // Toggle the group association
+        $video->toggleGroup($group);
+
+        $result = $group->hasGroupable($video)
+            ? __('Added to :group.', ['group' => $group->title])
+            : __('Removed from :group.', ['group' => $group->title]);
 
         Inertia::flash([
             'title' => (string) $video->name,
-            'description' => $group->hasGroupable($video)
-                ? __('Added to :group.', ['group' => $group->title])
-                : __('Removed from :group.', ['group' => $group->title]),
+            'description' => $result,
             'type' => 'success',
         ]);
