+++ src/App/Web/Notifications/Controllers/MarkAllNotificationsReadController.php
@@ -2,10 +2,8 @@
 
 declare(strict_types=1);
 
-namespace App\Api\Tags\Controllers;
+namespace App\Web\Notifications\Controllers;
 
-use Domain\Tags\Actions\SetTagsOrder;
-use Domain\Tags\Models\Tag;
 use Foundation\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -14,26 +12,25 @@ use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
-class TagOrderController extends Controller implements HasMiddleware
+class MarkAllNotificationsReadController extends Controller implements HasMiddleware
 {
     public static function middleware(): array
     {
         return [
             new Middleware('auth'),
             new Middleware('verified'),
-            new Middleware('role:super-admin'),
         ];
     }
 
     public function __invoke(Request $request): RedirectResponse
     {
-        Gate::authorize('create', Tag::class);
+        Gate::authorize('update', $request->user());
 
-        app(SetTagsOrder::class)->handle();
+        $request->user()->unreadNotifications()->update(['read_at' => now()]);
 
         Inertia::flash([
-            'title' => __('Tags reordered'),
-            'description' => __('The display order has been updated.'),
+            'title' => 'All caught up',
+            'description' => 'All notifications have been marked as read.',
             'type' => 'success',
         ]);
