+++ src/App/Web/Videos/Controllers/VideoDispatchTranscodeController.php
@@ -2,40 +2,40 @@
 
 declare(strict_types=1);
 
-namespace App\Api\Videos\Controllers;
+namespace App\Web\Videos\Controllers;
 
-use Domain\Videos\Actions\ProcessVideoImport;
+use Domain\Videos\Jobs\TranscodeVideo;
 use Domain\Videos\Models\Video;
 use Foundation\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
-class VideoImportController extends Controller implements HasMiddleware
+class VideoDispatchTranscodeController extends Controller implements HasMiddleware
 {
     public static function middleware(): array
     {
         return [
             new Middleware('auth'),
             new Middleware('verified'),
+            new Middleware('role:super-admin'),
         ];
     }
 
-    public function __invoke(Request $request): RedirectResponse
+    public function __invoke(Video $video): RedirectResponse
     {
-        Gate::authorize('create', Video::class);
+        Gate::authorize('update', $video);
 
-        app(ProcessVideoImport::class)->handle(
-            user: $request->user(),
-            disk: Video::getImportDisk()
+        TranscodeVideo::dispatchIf(
+            ! $video->hasTranscode(),
+            $video
         );
 
         Inertia::flash([
-            'title' => __('Video import initiated'),
-            'description' => __('Files are being processed in the background.'),
+            'title' => (string) $video->name,
+            'description' => __('Queued for transcoding.'),
             'type' => 'info',
         ]);
